+++ pkg/chunk/cache/redis_cache.go
@@ -3,7 +3,6 @@ package cache
 import (
 	"context"
 	"errors"
-	"flag"
 	"time"
 
 	"github.com/cortexproject/cortex/pkg/util"
@@ -29,15 +28,6 @@ type RedisCache struct {
 	requestDuration observableVecCollector
 }
 
-// RedisConfig defines how a RedisCache should be constructed.
-type RedisConfig struct {
-	Endpoint       string        `yaml:"endpoint,omitempty"`
-	Timeout        time.Duration `yaml:"timeout,omitempty"`
-	Expiration     time.Duration `yaml:"expiration,omitempty"`
-	MaxIdleConns   int           `yaml:"max_idle_conns,omitempty"`
-	MaxActiveConns int           `yaml:"max_active_conns,omitempty"`
-}
-
 type redisClient struct {
 	pool *redis.Pool
 }
@@ -47,24 +37,14 @@ var (
 		Namespace: "cortex",
 		Name:      "redis_request_duration_seconds",
 		Help:      "Total time spent in seconds doing redis requests.",
-		// Redis requests are very quick: smallest bucket is 16us, biggest is 1s
-		Buckets: prometheus.ExponentialBuckets(0.000016, 4, 8),
+		Buckets:   prometheus.ExponentialBuckets(0.00025, 4, 6),
 	}, []string{"method", "status_code", "name"})
 
 	errRedisQueryTimeout = errors.New("redis query timeout")
 )
 
-// RegisterFlagsWithPrefix adds the flags required to config this to the given FlagSet
-func (cfg *RedisConfig) RegisterFlagsWithPrefix(prefix, description string, f *flag.FlagSet) {
-	f.StringVar(&cfg.Endpoint, prefix+"redis.endpoint", "", description+"Redis service endpoint to use when caching chunks. If empty, no redis will be used.")
-	f.DurationVar(&cfg.Timeout, prefix+"redis.timeout", 100*time.Millisecond, description+"Maximum time to wait before giving up on redis requests.")
-	f.DurationVar(&cfg.Expiration, prefix+"redis.expiration", 0, description+"How long keys stay in the redis.")
-	f.IntVar(&cfg.MaxIdleConns, prefix+"redis.max-idle-conns", 80, description+"Maximum number of idle connections in pool.")
-	f.IntVar(&cfg.MaxActiveConns, prefix+"redis.max-active-conns", 0, description+"Maximum number of active connections in pool.")
-}
-
 // NewRedisCache creates a new RedisCache
-func NewRedisCache(cfg RedisConfig, name string, client RedisClient) *RedisCache {
+func NewRedisCache(cfg StoreConfig, name string, client RedisClient) *RedisCache {
 	// client != nil in unit tests
 	if client == nil {
 		client = &redisClient{
@@ -72,7 +52,7 @@ func NewRedisCache(cfg RedisConfig, name string, client RedisClient) *RedisCache
 				MaxIdle:   cfg.MaxIdleConns,
 				MaxActive: cfg.MaxActiveConns,
 				Dial: func() (redis.Conn, error) {
-					c, err := redis.Dial("tcp", cfg.Endpoint)
+					c, err := redis.Dial("tcp", cfg.Service)
 					if err != nil {
 						return nil, err
 					}
@@ -95,7 +75,7 @@ func NewRedisCache(cfg RedisConfig, name string, client RedisClient) *RedisCache
 	}
 
 	if err := cache.ping(context.Background()); err != nil {
-		level.Error(util.Logger).Log("msg", "error connecting to redis", "endpoint", cfg.Endpoint, "err", err)
+		level.Error(util.Logger).Log("msg", "error connecting to redis", "endpoint", cfg.Service, "err", err)
 	}
 
 	return cache